chore: SSL implementation on docker
Appended "certbot" service to our system, so that it will create SSL
certificates to be use by nginx. This cerficates are mapped on
./utils/certbot/www.

Created a base.conf file to redirect http request to https so that if
someone has the URL as http://domain can still use the website.

Modified solestyle.conf to serve content on 443 using SSL.