chore: CI/CD shopping_cart_management service
Created a shopping_cart_management.yml file to be run the jobs: build
and test when a PR is open to develop and then push_image when a commit
is pushed to develop. This workflow is only triggered if a file has been
updated inside the dir backend/shoppingCardManagement.

This has been done to include shopping_cart_management in the CI/CD
workflow.