feature: user authentication with Google
Updated frontend to show a Google button to either Sign-up or Sign-in to
the application. The label of this button is updated to reflect which
case.

Appended the Google provider to account_management, this provider uses
the idtoken library from google to perform a validation using the
IDTOKEN. The authentication is done directly to the server while the
frontend only controls the button.